feat: implement SDKCore_Collectors service for data collection and user management
