+++ packages/studiocms/src/sdk/effect/index.ts
@@ -3,3 +3,4 @@ export { SDKCore_FolderTree } from './foldertree.js';
 export { SDKCore_Generators } from './generators.js';
 export { SDKCore_Parsers } from './parsers.js';
 export { SDKCore_Users } from './users.js';
+export { SDKCore_Collectors } from './collectors.js';
